requirements.txt
# Django and Core Framework
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
python-dotenv==1.0.0

# File Processing (minimal)
python-docx==1.1.0
docx2txt==0.8

# AI Integration
openai==1.3.7
mistralai==0.0.12
requests==2.31.0

# Development and Utilities
whitenoise==6.6.0 